+++ kwimage/algo/_nms_backend/py_nms.py
@@ -1,11 +1,9 @@
-# -*- coding: utf-8 -*-
 """
 Fast R-CNN
 Copyright (c) 2015 Microsoft
 Licensed under The MIT License [see LICENSE for details]
 Written by Ross Girshick
 """
-from __future__ import absolute_import, division, print_function, unicode_literals
 import numpy as np
 import warnings